+++ Missile.cpp
@@ -79,8 +79,8 @@ void Missile::UpdatePhysics(Game& game)
 void Missile::CalculateCollider(Vector2 cameraOffset)
 {
 	// set the collision bounds position to where the entity actually is
-	collisionBounds->x = position.x + collider->x - cameraOffset.x;
-	collisionBounds->y = position.y + collider->y - cameraOffset.y;
+	collisionBounds->x = (int)(position.x + collider->x - cameraOffset.x);
+	collisionBounds->y = (int)(position.y + collider->y - cameraOffset.y);
 
 	// scale the bounds of the sprite by a number to set the collider's width and height
 	collisionBounds->w = colliderWidth;
